Enable additional g-code commands
Enable M500 and M501 to allow settings like bed leveling data to
be saved to EEPROM memory. Note that power cycling the printer will
not clear any changes saved with M500, if required defaults can be
loaded again with M502 which must be saved with M500.
Enable G12 for nozzle cleaning which may be useful to run prior to
starting a print to prime the nozzle.